requirements.txt
click>=6.7,6.8
cloudshell-automation-api>=8.3,<8.4
cloudshell-rest-api>=8.2,<8.3
pysmb>=1.1,<1.2
PyYAML>=3.12,<3.13
requests>=2.18,<2.19
